use-internal-oauth-proxy-image will only apply to pods that have an oauth-proxy container already present
Currently the use-internal-oauth-proxy-image policy will apply to all pods even if they don't have an oauth-proxy container.
This leads to issues with pods created from sources other than RHOAI since it tries to pull the oauth-proxy image when it
should not be. This PR will make it so this policy only applies to pods with an oauth-proxy container already present.

Signed-off-by: Isaiah Stapleton <istaplet@redhat.com>