Remove validate-cs210 policy (#17)
Students from cs210 are having performance issues and are going to test
using Small resource size instead of XSmall. So removing this policy for
now so they can test that, and then will add it back with the proper
resource size once we figure out if that solves the issue.

Signed-off-by: Isaiah Stapleton <istaplet@redhat.com>